Added rulebook and refined menu
I have refined the menu otpions to remove the 'Reset' option and put it
in the resources section as a button. Additionally, I finally found a
way to display the Pdfs (albeit slowly) in-app. Navigation is taking
shape, as well. Need to find a way to unify the navigation under a
single structure...inherited Activity type?