Added card search
Replaced the deckbox with a card search (for now), which uses the search
strings from carddb.com to display the carddb search results in a
webview control. I would like to limit the return to only the section
with the results, but it looks like the EvaluateJavascript method is not
retrieving appropriate results, or the results are not formatted
correctly.